+++ src/jobs.jl
@@ -42,7 +42,10 @@ end
  - `ncpu::Int64 = 1`: number of CPU this job is about to use.
  - `mem::Int64 = 0`: number of memory this job is about to use (supports TB, GB, MB, KB, B=1).
  - `schedule_time::Union{DateTime,Period} = DateTime(0)`: The expected time to run.
- - `dependency::Vector{Pair{Symbol,Int64}}`: defer job until specified jobs reach specified state (QUEUING, RUNNING, DONE, FAILED, CANCELLED, PAST). PAST is the super set of DONE, FAILED, CANCELLED, which means the job will not run in the future.
+ - `dependency`: defer job until specified jobs reach specified state (QUEUING, RUNNING, DONE, FAILED, CANCELLED, PAST). PAST is the super set of DONE, FAILED, CANCELLED, which means the job will not run in the future. Eg: `DONE => job`, `[DONE => job1; PAST => job2]`.
+
+ > To be compatible with old versions, you can also use job id (Int): `[DONE => job.id]`.
+
  - `wall_time::Period = Week(1)`: wall clock time limit.
  - `priority::Int = 20`: lower means higher priority.
 
@@ -68,23 +71,24 @@ mutable struct Job
     wall_time::Period
     state::Symbol
     priority::Int
-    dependency::Vector{Pair{Symbol,Int64}}
+    dependency::Vector{Pair{Symbol,Union{Int64, Job}}}
     task::Union{Task,Nothing}
     stdout_file::String
     stderr_file::String
     _thread_id::Int
 
     # original
-    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::DateTime, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Vector{Pair{Symbol,Int64}}, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String)
+    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::DateTime, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Vector, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String)
         if !(typeof(wall_time) <: Union{Millisecond,Second,Minute,Hour,Day,Week})
             error("Job.wall_time is not one of Union{Millisecond,Second,Minute,Hour,Day,Week}")
         end
         check_ncpu_mem(ncpu, mem)
+        dependency = convert(Vector{Pair{Symbol,Union{Int64, Job}}}, dependency)
         new(id, name, user, ncpu, mem, schedule_time, create_time, start_time, stop_time, wall_time, state, priority, dependency, task, stdout_file, stderr_file, 0)
     end
 
     # schedule_time::Period
-    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::Period, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Vector{Pair{Symbol,Int64}}, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String)
+    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::Period, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Vector, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String)
         if !(typeof(wall_time) <: Union{Millisecond,Second,Minute,Hour,Day,Week})
             error("Job.wall_time is not one of Union{Millisecond,Second,Minute,Hour,Day,Week}")
         end
@@ -92,21 +96,23 @@ mutable struct Job
             error("Job.schedule_time is not one of Union{DateTime,Millisecond,Second,Minute,Hour,Day,Week}")
         end
         check_ncpu_mem(ncpu, mem)
+        dependency = convert(Vector{Pair{Symbol,Union{Int64, Job}}}, dependency)
         new(id, name, user, ncpu, mem, now() + schedule_time, create_time, start_time, stop_time, wall_time, state, priority, dependency, task, stdout_file, stderr_file, 0)
     end
 
     # dependency::Pair{Symbol,Int64}
-    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::DateTime, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Pair{Symbol,Int64}, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String)
+    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::DateTime, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Pair{Symbol,T}, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String) where T
         if !(typeof(wall_time) <: Union{Millisecond,Second,Minute,Hour,Day,Week})
             error("Job.wall_time is not one of Union{Millisecond,Second,Minute,Hour,Day,Week}")
         end
         check_ncpu_mem(ncpu, mem)
-        new(id, name, user, ncpu, mem, schedule_time, create_time, start_time, stop_time, wall_time, state, priority, [dependency], task, stdout_file, stderr_file, 0)
+        dependency = Pair{Symbol,Union{Int64, Job}}[dependency]
+        new(id, name, user, ncpu, mem, schedule_time, create_time, start_time, stop_time, wall_time, state, priority, dependency, task, stdout_file, stderr_file, 0)
     end
 
     # schedule_time::Period
     # dependency::Pair{Symbol,Int64}
-    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::Period, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Pair{Symbol,Int64}, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String)
+    function Job(id::Int64, name::String, user::String, ncpu::Int64, mem::Int64, schedule_time::Period, create_time::DateTime, start_time::DateTime, stop_time::DateTime, wall_time::Period, state::Symbol, priority::Int, dependency::Pair{Symbol,T}, task::Union{Task,Nothing}, stdout_file::String, stderr_file::String) where T
         if !(typeof(wall_time) <: Union{Millisecond,Second,Minute,Hour,Day,Week})
             error("Job.wall_time is not one of Union{Millisecond,Second,Minute,Hour,Day,Week}")
         end
@@ -114,7 +120,8 @@ mutable struct Job
             error("Job.schedule_time is not one of Union{DateTime,Millisecond,Second,Minute,Hour,Day,Week}")
         end
         check_ncpu_mem(ncpu, mem)
-        new(id, name, user, ncpu, mem, now() + schedule_time, create_time, start_time, stop_time, wall_time, state, priority, [dependency], task, stdout_file, stderr_file, 0)
+        dependency = Pair{Symbol,Union{Int64, Job}}[dependency]
+        new(id, name, user, ncpu, mem, now() + schedule_time, create_time, start_time, stop_time, wall_time, state, priority, dependency, task, stdout_file, stderr_file, 0)
     end
 end